No NaNs.

+++ src/main/scala/Engine.scala
@@ -5,7 +5,7 @@ import io.prediction.controller.Engine
 
 case class Query(content: String) extends Serializable
 
-case class PredictedResult(sentiments: List[Integer]) extends Serializable
+case class PredictedResult(sentiments: List[Double]) extends Serializable
 
 object VanillaEngine extends IEngineFactory {
   def apply() = {
